feat[set up of html and css files]
initiated set up of html by using boiler plate and container plus
all items. In css file used grid to design fundamentals of the layout.
This will visually separate the contents of the website for the user.
Design so far is responsive to different screen sizes and resizing of the
browser window.